+++ free-categories.cabal
@@ -16,6 +16,9 @@ extra-source-files:  CHANGELOG.md, README.md
 
 library
   exposed-modules:     Control.Category.Free
+                       Data.Quiver
+                       Data.Quiver.Bifunctor
+                       Data.Quiver.Functor
   build-depends:       base >=4.12 && <=5
   hs-source-dirs:      src
   default-language:    Haskell2010
