+++ src/Control/Category/Free.hs
@@ -37,11 +37,8 @@ from quivers to `Category`s may be defined up to isomorphism as
 -}
 
 {-# LANGUAGE
-    FlexibleInstances
-  , FunctionalDependencies
-  , GADTs
+    GADTs
   , LambdaCase
-  , MultiParamTypeClasses
   , PatternSynonyms
   , PolyKinds
   , QuantifiedConstraints
@@ -54,38 +51,15 @@ module Control.Category.Free
   , pattern (:<<)
   , FoldPath (..)
   , Category (..)
-  , CFunctor (..)
-  , CFoldable (..)
-  , CTraversable (..)
-  , CPointed (..)
-  , CStrong (..)
-  , CApplicative (..)
-  , CMonad (..)
   , CFree (..)
   , toPath
   , creverse
   , beforeAll
   , afterAll
-  , KQ (..)
-  , ProductQ (..)
-  , assocQ
-  , disassocQ
-  , productQ
-  , swapQ
-  , Quiver (..)
-  , ApQ (..)
-  , OpQ (..)
-  , IsoQ (..)
-  , IQ (..)
-  , ComposeQ (..)
-  , ExtendQ (..)
-  , LiftQ (..)
-  , CBifunctor
-  , CProfunctor
-  , CMonoidal
-  , CClosed
   ) where
 
+import Data.Quiver
+import Data.Quiver.Functor
 import Control.Category
 import Control.Monad (join)
 import Prelude hiding (id, (.))
@@ -154,141 +128,6 @@ instance CPointed FoldPath where csingleton p = FoldPath $ \ k -> k p
 instance CMonad FoldPath where cjoin (FoldPath f) = f id
 instance CFree FoldPath
 
-{- | An endfunctor of quivers.
-
-prop> cmap id = id
-prop> cmap (g . f) = cmap g . cmap f
-
-A functor from quivers to `Category`s
-has @(CFunctor c, forall p. Category (c p))@ with
-
-prop> cmap f id = id
-prop> cmap f (q . p) = cmap f q . cmap f p
--}
-class CFunctor c where
-  cmap :: (forall x y. p x y -> q x y) -> c p x y -> c q x y
-
-{- | Generalizing `Foldable` from `Monoid`s to `Category`s.
-
-prop> cmap f = cfoldMap (csingleton . f)
--}
-class CFunctor c => CFoldable c where
-  {- | Map each element of the structure to a `Category`,
-  and combine the results.-}
-  cfoldMap :: Category q => (forall x y. p x y -> q x y) -> c p x y -> q x y
-  {- | Combine the elements of a structure using a `Category`.-}
-  cfold :: Category q => c q x y -> q x y
-  cfold = cfoldMap id
-  {- | Right-associative fold of a structure.
-
-  In the case of `Path`s, `cfoldr`, when applied to a binary operator,
-  a starting value, and a `Path`, reduces the `Path` using the binary operator,
-  from right to left:
-
-  prop> cfoldr (?) q (p1 :>> p2 :>> ... :>> pn :>> Done) == p1 ? (p2 ? ... (pn ? q) ...)
-  -}
-  cfoldr :: (forall x y z . p x y -> q y z -> q x z) -> q y z -> c p x y -> q x z
-  cfoldr (?) q c = getLiftQ (cfoldMap (\ x -> LiftQ (\ y -> x ? y)) c) q
-  {- | Left-associative fold of a structure.
-
-  In the case of `Path`s, `cfoldl`, when applied to a binary operator,
-  a starting value, and a `Path`, reduces the `Path` using the binary operator,
-  from left to right:
-
-  prop> cfoldl (?) q (p1 :>> p2 :>> ... :>> pn :>> Done) == (... ((q ? p1) ? p2) ? ...) ? pn
-  -}
-  cfoldl :: (forall x y z . q x y -> p y z -> q x z) -> q x y -> c p y z -> q x z
-  cfoldl (?) q c = getExtendQ (cfoldMap (\ x -> ExtendQ (\ y -> y ? x)) c) q
-  {- | Map each element of the structure to a `Monoid`,
-  and combine the results.-}
-  ctoMonoid :: Monoid m => (forall x y. p x y -> m) -> c p x y -> m
-  ctoMonoid f = getKQ . cfoldMap (KQ . f)
-  {- | Map each element of the structure, and combine the results in a list.-}
-  ctoList :: (forall x y. p x y -> a) -> c p x y -> [a]
-  ctoList f = ctoMonoid (pure . f)
-  {- | Map each element of a structure to an `Applicative` on a `Category`,
-  evaluate from left to right, and combine the results.-}
-  ctraverse_
-    :: (Applicative m, Category q)
-    => (forall x y. p x y -> m (q x y)) -> c p x y -> m (q x y)
-  ctraverse_ f = getApQ . cfoldMap (ApQ . f)
-
-{- | Generalizing `Traversable` to quivers.-}
-class CFoldable c => CTraversable c where
-  {- | Map each element of a structure to an `Applicative` on a quiver,
-  evaluate from left to right, and collect the results.-}
-  ctraverse
-    :: Applicative m
-    => (forall x y. p x y -> m (q x y)) -> c p x y -> m (c q x y)
-
-{- | Embed a single quiver arrow with `csingleton`.-}
-class CPointed c where csingleton :: p x y -> c p x y
-
-{- | Strength for quiver endofunctors with respect to `ProductQ`.
-
-/Note:/ Every `Functor` is strong with respect to @(,)@,
-but not every `CFunctor` is strong with respect to @ProductQ@.
-
-prop> csecond . productQ id csecond . disassocQ = cmap disassocQ . csecond
-prop> cmap sndQ . csecond = sndQ
-
-`cfirst` and `csecond` are related as
-prop> cfirst = cmap swapQ . csecond . swapQ
-prop> csecond = cmap swapQ . cfirst . swapQ
--}
-class CFunctor c => CStrong c where
-  cfirst :: ProductQ (c p) q x y -> c (ProductQ p q) x y
-  cfirst = cmap swapQ . csecond . swapQ
-  csecond :: ProductQ p (c q) x y -> c (ProductQ p q) x y
-  csecond = cmap swapQ . cfirst . swapQ
-  {-# MINIMAL cfirst | csecond #-}
-
-{- | Generalize `Applicative` to quivers.
-
-The laws of a strong lax monoidal endofunctor hold.
-
->>> let cunit = csingleton (KQ ())
->>> let ctimes = czip ProductQ
-
-prop> cmap (f `productQ` g) (p `ctimes` q) = cmap f p `ctimes` cmap g q
-prop> cmap sndQ (cunit `ctimes` q) = q
-prop> cmap fstQ (p `ctimes` cunit) = p
-prop> cmap assocQ (p `ctimes` (q `ctimes` r)) = (p `ctimes` q) `ctimes` r
-
-The functions `cap` and `czip` are related as
-
-prop> cap = czip getQuiver
-prop> czip f p q = (Quiver . f) `cmap` p `cap` q
--}
-class (CStrong c, CPointed c) => CApplicative c where
-  cap :: c (Quiver p q) x y -> c p x y -> c q x y
-  cap = czip getQuiver
-  czip
-    :: (forall x y. p x y -> q x y -> r x y)
-    -> c p x y -> c q x y -> c r x y
-  czip f p q = (Quiver . f) `cmap` p `cap` q
-  {-# MINIMAL cap | czip #-}
-
-{- | Generalize `Monad` to quivers.
-
-Associativity and left and right identity laws hold.
-
-prop> cjoin . cjoin = cjoin . cmap cjoin
-prop> cjoin . csingleton = id
-prop> cjoin . cmap csingleton = id
-
-The functions `cbind` and `cjoin` are related as
-
-prop> cjoin = cbind id
-prop> cbind f p = cjoin (cmap f p)
--}
-class (CFunctor c, CPointed c) => CMonad c where
-  cjoin :: c (c p) x y -> c p x y
-  cjoin = cbind id
-  cbind :: (forall x y. p x y -> c q x y) -> c p x y -> c q x y
-  cbind f p = cjoin (cmap f p)
-  {-# MINIMAL cjoin | cbind #-}
-
 {- | Unpacking the definition of a left adjoint to the forgetful functor
 from `Category`s to quivers, any
 
@@ -326,221 +165,3 @@ afterAll
   :: (CFoldable c, CFree path)
   => (forall x. p x x) -> c p x y -> path p x y
 afterAll sep = cfoldMap (\p -> csingleton p >>> csingleton sep)
-
-{- | Turn a `Monoid` into a `Category`,
-used in the default definition of `ctoMonoid`.-}
-newtype KQ r x y = KQ {getKQ :: r} deriving (Eq, Ord, Show)
-instance (Semigroup m, x ~ y) => Semigroup (KQ m x y) where
-  KQ f <> KQ g = KQ (f <> g)
-instance (Monoid m, x ~ y) => Monoid (KQ m x y) where mempty = id
-instance Monoid m => Category (KQ m) where
-  id = KQ mempty
-  KQ g . KQ f = KQ (f <> g)
-
-{- | Product of quivers.-}
-data ProductQ p q x y = ProductQ
-  { fstQ :: p x y
-  , sndQ :: q x y
-  } deriving (Eq, Ord, Show)
-instance (Category p, Category q, x ~ y)
-  => Semigroup (ProductQ p q x y) where (<>) = (>>>)
-instance (Category p, Category q, x ~ y)
-  => Monoid (ProductQ p q x y) where mempty = id
-instance (Category p, Category q) => Category (ProductQ p q) where
-  id = ProductQ id id
-  ProductQ pyz qyz . ProductQ pxy qxy = ProductQ (pyz . pxy) (qyz . qxy)
-instance CFunctor (ProductQ p) where cmap f (ProductQ p q) = ProductQ p (f q)
-instance CFoldable (ProductQ p) where cfoldMap f (ProductQ _ q) = f q
-instance CTraversable (ProductQ p) where
-  ctraverse f (ProductQ p q) = ProductQ p <$> f q
-instance CBifunctor ProductQ where
-  cbimap f g (ProductQ p q) = ProductQ (f p) (g q)
-
-{- | Associator of `ProductQ`.-}
-assocQ :: ProductQ p (ProductQ q r) x y -> ProductQ (ProductQ p q) r x y
-assocQ (ProductQ p (ProductQ q r)) = ProductQ (ProductQ p q) r
-
-{- | Inverse associator of `ProductQ`.-}
-disassocQ :: ProductQ (ProductQ p q) r x y -> ProductQ p (ProductQ q r) x y
-disassocQ (ProductQ (ProductQ p q) r) = ProductQ p (ProductQ q r)
-
-{- | Map over the `fstQ` and `sndQ` of a `ProductQ`.-}
-productQ
-  :: (p0 x y -> p1 x y)
-  -> (q0 x y -> q1 x y)
-  -> ProductQ p0 q0 x y
-  -> ProductQ p1 q1 x y
-productQ f g (ProductQ p q) = ProductQ (f p) (g q)
-
-{- | Symmetry of `ProductQ`.-}
-swapQ :: ProductQ p q x y -> ProductQ q p x y
-swapQ (ProductQ p q) = ProductQ q p
-
-{- | Morphism components of quivers.
-
-Quivers form a Cartesian closed category with
-  * product `ProductQ`
-  * unit `KQ ()`
-  * internal hom `Quiver`
--}
-newtype Quiver p q x y = Quiver { getQuiver :: p x y -> q x y }
-instance CFunctor (Quiver p) where cmap g (Quiver f) = Quiver (g . f)
-instance CPointed (Quiver p) where csingleton q = Quiver (const q)
-instance CStrong (Quiver p) where
-  csecond (ProductQ p (Quiver f)) = Quiver (ProductQ p . f)
-instance CApplicative (Quiver p) where
-  cap (Quiver cf) (Quiver cq) = Quiver (\p -> getQuiver (cf p) (cq p))
-instance CMonad (Quiver p) where
-  cjoin (Quiver q) = Quiver (\p -> getQuiver (q p) p)
-instance CProfunctor Quiver where
-  cdimap f h (Quiver g) = Quiver (h . g . f)
-
-{- | Turn an `Applicative` over a `Category` into a `Category`,
-used in the default definition of `ctraverse_`.-}
-newtype ApQ m c x y = ApQ {getApQ :: m (c x y)} deriving (Eq, Ord, Show)
-instance (Applicative m, Category c, x ~ y)
-  => Semigroup (ApQ m c x y) where (<>) = (>>>)
-instance (Applicative m, Category c, x ~ y)
-  => Monoid (ApQ m c x y) where mempty = id
-instance (Applicative m, Category c) => Category (ApQ m c) where
-  id = ApQ (pure id)
-  ApQ g . ApQ f = ApQ ((.) <$> g <*> f)
-instance Functor t => CFunctor (ApQ t) where
-  cmap f (ApQ t) = ApQ (f <$> t)
-instance Applicative t => CPointed (ApQ t) where
-  csingleton = ApQ . pure
-instance Functor t => CStrong (ApQ t) where
-  csecond (ProductQ p (ApQ t)) = ApQ (ProductQ p <$> t)
-instance Applicative t => CApplicative (ApQ t) where
-  czip f (ApQ tp) (ApQ tq) = ApQ (f <$> tp <*> tq)
-instance Monad t => CMonad (ApQ t) where
-  cbind f (ApQ t) = ApQ $ do
-    p <- t
-    getApQ $ f p
-
-{- | Reverse all the arrows in a quiver.-}
-newtype OpQ c x y = OpQ {getOpQ :: c y x} deriving (Eq, Ord, Show)
-instance (Category c, x ~ y) => Semigroup (OpQ c x y) where (<>) = (>>>)
-instance (Category c, x ~ y) => Monoid (OpQ c x y) where mempty = id
-instance Category c => Category (OpQ c) where
-  id = OpQ id
-  OpQ g . OpQ f = OpQ (f . g)
-instance CFunctor OpQ where cmap f = OpQ . f . getOpQ
-
-{- | Turn all arrows in a quiver into bidirectional edges.-}
-data IsoQ c x y = IsoQ
-  { up :: c x y
-  , down :: c y x
-  } deriving (Eq, Ord, Show)
-instance (Category c, x ~ y) => Semigroup (IsoQ c x y) where (<>) = (>>>)
-instance (Category c, x ~ y) => Monoid (IsoQ c x y) where mempty = id
-instance Category c => Category (IsoQ c) where
-  id = IsoQ id id
-  (IsoQ yz zy) . (IsoQ xy yx) = IsoQ (yz . xy) (yx . zy)
-instance CFunctor IsoQ where
-  cmap f (IsoQ u d) = IsoQ (f u) (f d)
-
-{- | The identity functor on quivers. -}
-newtype IQ c x y = IQ {getIQ :: c x y} deriving (Eq, Ord, Show)
-instance (Category c, x ~ y) => Semigroup (IQ c x y) where (<>) = (>>>)
-instance (Category c, x ~ y) => Monoid (IQ c x y) where mempty = id
-instance Category c => Category (IQ c) where
-  id = IQ id
-  IQ g . IQ f = IQ (g . f)
-instance CFunctor IQ where cmap f = IQ . f . getIQ
-instance CFoldable IQ where cfoldMap f (IQ c) = f c
-instance CTraversable IQ where ctraverse f (IQ c) = IQ <$> f c
-instance CPointed IQ where csingleton = IQ
-instance CStrong IQ where csecond (ProductQ p (IQ q)) = IQ (ProductQ p q)
-instance CApplicative IQ where czip f (IQ p) (IQ q) = IQ (f p q)
-instance CMonad IQ where cjoin = getIQ
-
-data ReflQ r x y where ReflQ :: r -> ReflQ r x x
-instance Monoid m => Category (ReflQ m) where
-  id = ReflQ mempty
-  ReflQ yz . ReflQ xy = ReflQ (xy <> yz)
-
-data ComposeQ p q x z = forall y. ComposeQ (p y z) (q x y)
-deriving instance (forall x y. (Show (p x y), Show (q x y)))
-  => Show (ComposeQ p q x y)
-instance CFunctor (ComposeQ p) where
-  cmap f (ComposeQ p q) = ComposeQ p (f q)
-instance Category p => CPointed (ComposeQ p) where
-  csingleton = ComposeQ id
-instance Category p => CMonad (ComposeQ p) where
-  cjoin (ComposeQ yz (ComposeQ xy q)) = ComposeQ (yz . xy) q
-instance CBifunctor ComposeQ where
-  cbimap f g (ComposeQ p q) = ComposeQ (f p) (g q)
-
-newtype ExtendQ p q x y = ExtendQ {getExtendQ :: forall w. p w x -> q w y}
-instance CFunctor (ExtendQ p) where
-  cmap g (ExtendQ f) = ExtendQ (g . f)
-instance (p ~ q, x ~ y) => Semigroup (ExtendQ p q x y) where (<>) = (>>>)
-instance (p ~ q, x ~ y) => Monoid (ExtendQ p q x y) where mempty = id
-instance p ~ q => Category (ExtendQ p q) where
-  id = ExtendQ id
-  ExtendQ g . ExtendQ f = ExtendQ (g . f)
-instance CProfunctor ExtendQ where
-  cdimap f h (ExtendQ g) = ExtendQ (h . g . f)
-
-newtype LiftQ p q x y = LiftQ {getLiftQ :: forall z. p y z -> q x z}
-instance CFunctor (LiftQ p) where
-  cmap g (LiftQ f) = LiftQ (g . f)
-instance (p ~ q, x ~ y) => Semigroup (LiftQ p q x y) where (<>) = (>>>)
-instance (p ~ q, x ~ y) => Monoid (LiftQ p q x y) where mempty = id
-instance p ~ q => Category (LiftQ p q) where
-  id = LiftQ id
-  LiftQ f . LiftQ g = LiftQ (g . f)
-instance CProfunctor LiftQ where
-  cdimap f h (LiftQ g) = LiftQ (h . g . f)
-
-class (CBifunctor prod, CProfunctor lhom, CProfunctor rhom)
-  => CClosed prod lhom rhom | prod -> lhom, prod -> rhom where
-    ccurry :: (forall x y. prod p q x y -> r x y) -> p x y -> lhom q r x y
-    cuncurry :: (forall x y. p x y -> lhom q r x y) -> prod p q x y -> r x y
-    cflurry :: (forall x y. prod p q x y -> r x y) -> q x y -> rhom p r x y
-    cunflurry :: (forall x y. q x y -> rhom p r x y) -> prod p q x y -> r x y
-instance CClosed ProductQ Quiver Quiver where
-  ccurry f p = Quiver (\q -> f (ProductQ p q))
-  cuncurry f (ProductQ p q) = getQuiver (f p) q
-  cflurry f q = Quiver (\p -> f (ProductQ p q))
-  cunflurry f (ProductQ p q) = getQuiver (f q) p
-instance CClosed ComposeQ ExtendQ LiftQ where
-  ccurry f p = ExtendQ (\q -> f (ComposeQ p q))
-  cuncurry f (ComposeQ p q) = getExtendQ (f p) q
-  cflurry f q = LiftQ (\p -> f (ComposeQ p q))
-  cunflurry f (ComposeQ p q) = getLiftQ (f q) p
-
-class CBifunctor prod => CMonoidal prod unit | prod -> unit where
-  cintro1 :: p x y -> prod unit p x y
-  cintro2 :: p x y -> prod p unit x y
-  celim1 :: prod unit p x y -> p x y
-  celim2 :: prod p unit x y -> p x y
-  cassoc :: prod p (prod q r) x y -> prod (prod p q) r x y
-  cdisassoc :: prod (prod p q) r x y -> prod p (prod q r) x y
-instance CMonoidal ProductQ (KQ ()) where
-  cintro1 p = ProductQ (KQ ()) p
-  cintro2 p = ProductQ p (KQ ())
-  celim1 (ProductQ _ p) = p
-  celim2 (ProductQ p _) = p
-  cassoc (ProductQ p (ProductQ q r)) = ProductQ (ProductQ p q) r
-  cdisassoc (ProductQ (ProductQ p q) r) = ProductQ p (ProductQ q r)
-instance CMonoidal ComposeQ (ReflQ ()) where
-  cintro1 p = ComposeQ (ReflQ ()) p
-  cintro2 p = ComposeQ p (ReflQ ())
-  celim1 (ComposeQ (ReflQ ()) p) = p
-  celim2 (ComposeQ p (ReflQ ())) = p
-  cassoc (ComposeQ p (ComposeQ q r)) = ComposeQ (ComposeQ p q) r
-  cdisassoc (ComposeQ (ComposeQ p q) r) = ComposeQ p (ComposeQ q r)
-
-class (forall q. CFunctor (prod q)) => CBifunctor prod where
-  cbimap
-    :: (forall x y. p x y -> p' x y)
-    -> (forall x y. q x y -> q' x y)
-    -> prod p q x y -> prod p' q' x y
-
-class (forall q. CFunctor (hom q)) => CProfunctor hom where
-  cdimap
-    :: (forall x y. p' x y -> p x y)
-    -> (forall x y. q x y -> q' x y)
-    -> hom p q x y -> hom p' q' x y
